Adding in 290 world vak 245 (#109)
* adding in 290 world

* trying to add in toon shader

* add toon shader

* auto texture mipmap level calculation

* Update Camera.hpp

* change skybox

* make it happy

* add new models

* Update sandbox.lua

* have digi twin on tree stump

* make material colors vec4

* continue with transparency

* something to look into in the future

* adding more colliders

* fix 290 world transparency issue

* moar colliders

* more colliders + rotation fix

* add more colliders and barriers

* update math functions

* no more rp3d crash

* last colliders and touchups

* clangify

---------

+++ Core/src/Physics/include/Physics/PhysicsEngine.hpp
@@ -53,7 +53,7 @@ namespace Vakol
          * @param radius of sphere
          * @return SphereCollider
          */
-        SphereCollider CreateSphereCollider(double radius);
+        SphereCollider CreateSphereCollider(float radius);
 
         /**
          * @brief Create a Capsule Collider object
@@ -62,7 +62,7 @@ namespace Vakol
          * @param height of capsule
          * @return CapsuleCollider
          */
-        CapsuleCollider CreateCapsuleCollider(double radius, double height);
+        CapsuleCollider CreateCapsuleCollider(float radius, float height);
 
         /**
          * @brief Create a Mesh Collider object
